Remove InitAndStartAgentDemultiplexerTest
