+++ cmd/agent/subcommands/run/internal/settings/runtime_settings_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/dogstatsd"
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/server"
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug"
+	"github.com/DataDog/datadog-agent/comp/forwarder/defaultforwarder"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/fx"
@@ -33,7 +34,8 @@ func TestDogstatsdMetricsStats(t *testing.T) {
 
 	opts := aggregator.DefaultAgentDemultiplexerOptions()
 	opts.DontStartForwarders = true
-	demux := aggregator.InitAndStartAgentDemultiplexerTest(opts, "hostname")
+	forwarder := fxutil.Test[defaultforwarder.Component](t, defaultforwarder.MockModule)
+	demux := aggregator.InitAndStartAgentDemultiplexer(forwarder, opts, "hostname")
 
 	deps := fxutil.Test[testDeps](t, fx.Options(
 		core.MockBundle,
