+++ comp/forwarder/defaultforwarder/component.go
@@ -26,3 +26,13 @@ type Component interface {
 var Module = fxutil.Component(
 	fx.Provide(newForwarder),
 )
+
+// Mock implements mock-specific methods.
+type Mock interface {
+	Component
+}
+
+// MockModule defines the fx options for the mock component.
+var MockModule = fxutil.Component(
+	fx.Provide(newMockForwarder),
+)
