+++ pkg/compliance/agent/agent_test.go
@@ -14,6 +14,7 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/DataDog/datadog-agent/comp/forwarder/defaultforwarder"
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/DataDog/datadog-agent/pkg/compliance/checks"
@@ -21,6 +22,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/compliance/mocks"
 	"github.com/DataDog/datadog-agent/pkg/logs/config"
 	"github.com/DataDog/datadog-agent/pkg/util"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -97,7 +99,8 @@ func TestRunK8s(t *testing.T) {
 
 	opts := aggregator.DefaultAgentDemultiplexerOptions()
 	opts.DontStartForwarders = true
-	aggregator.InitAndStartAgentDemultiplexerTest(opts, "foo")
+	forwarder := fxutil.Test[defaultforwarder.Component](t, defaultforwarder.MockModule)
+	aggregator.InitAndStartAgentDemultiplexer(forwarder, opts, "foo")
 
 	e := enterTempEnv(t, true)
 	defer e.leave()
@@ -173,7 +176,8 @@ func TestRunDocker(t *testing.T) {
 
 	opts := aggregator.DefaultAgentDemultiplexerOptions()
 	opts.DontStartForwarders = true
-	aggregator.InitAndStartAgentDemultiplexerTest(opts, "foo")
+	forwarder := fxutil.Test[defaultforwarder.Component](t, defaultforwarder.MockModule)
+	aggregator.InitAndStartAgentDemultiplexer(forwarder, opts, "foo")
 
 	e := enterTempEnv(t, false)
 	defer e.leave()
@@ -266,7 +270,8 @@ func TestRunChecks(t *testing.T) {
 
 	opts := aggregator.DefaultAgentDemultiplexerOptions()
 	opts.DontStartForwarders = true
-	aggregator.InitAndStartAgentDemultiplexerTest(opts, "foo")
+	forwarder := fxutil.Test[defaultforwarder.Component](t, defaultforwarder.MockModule)
+	aggregator.InitAndStartAgentDemultiplexer(forwarder, opts, "foo")
 
 	e := enterTempEnv(t, false)
 	defer e.leave()
