+++ pkg/metadata/scheduler_test.go
@@ -12,8 +12,11 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/DataDog/datadog-agent/comp/forwarder/defaultforwarder"
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
+	"github.com/DataDog/datadog-agent/pkg/forwarder"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
 type MockCollector struct {
@@ -56,7 +59,8 @@ func TestNewScheduler(t *testing.T) {
 
 	opts := aggregator.DefaultAgentDemultiplexerOptions()
 	opts.DontStartForwarders = true
-	demux := aggregator.InitAndStartAgentDemultiplexerTest(opts, "hostname")
+	forwarder := fxutil.Test[defaultforwarder.Component](t, defaultforwarder.MockModule)
+	demux := aggregator.InitAndStartAgentDemultiplexer(forwarder, opts, "hostname")
 	c := NewScheduler(demux)
 
 	assert.Equal(t, demux, c.demux)
@@ -195,6 +199,7 @@ func TestTriggerAndResetCollectorTimer(t *testing.T) {
 func buildDemultiplexer() aggregator.Demultiplexer {
 	opts := aggregator.DefaultAgentDemultiplexerOptions()
 	opts.DontStartForwarders = true
-	demux := aggregator.InitAndStartAgentDemultiplexerTest(opts, "hostname")
+	forwarder := forwarder.NewDefaultForwarder(forwarder.NewOptions(nil))
+	demux := aggregator.InitAndStartAgentDemultiplexer(forwarder, opts, "hostname")
 	return demux
 }
